refactor: inline session runtime from playbooks into server
The playbook session-runtime adapter layer (createPlaybookSessionRuntime,
pickPlaybookAgent, PlaybookAgentIdentityMismatchError, and ~15 input/output
types) lived in @openagentpack/playbooks but had exactly one production
consumer: the server's playbook-session-adapter. This commit:

- Moves the runtime orchestration and agent-pick logic into a new
  apps/server/.../playbook-session-adapter/runtime.ts with all generic
  type parameters instantiated to their concrete server types.
- Replaces the four adapter interfaces with a single concrete
  PlaybookSessionRuntimeDeps type (dependency injection preserved for
  testability, but no cross-package abstraction).
- Migrates tests to apps/server with type-safe fake constructors.
- Removes session-runtime.ts from @openagentpack/playbooks and cleans
  up all re-exports from the package index.
- Eliminates the duplicate ListPlaybookSessionsInput definition in
  sessions.ts (now imported from runtime.ts).

Behavioral changes: none. The PlaybookAgentIdentityMismatchError still
surfaces as HTTP 500 (no route-level catch); a future PR can map it to
4xx if desired.

Co-developed-by: OpenCode <noreply@opencode.ai>
